refactor(doctor): isolate resource classification
